requirement.txt
asgiref==3.4.1
autopep8==1.6.0
Django==4.0.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
psycopg2-binary==2.9.3
pycodestyle==2.8.0
PyJWT==2.3.0
pymemcache==3.5.0
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
